[FLINK-4613] Minor fixes in IALS
